requirements.txt
# pip install -r requirements.txt

absl-py @ file:///C:/b/abs_5babsu7y5x/croot/absl-py_1666362945682/work
accelerate==0.17.1
aiohttp @ file:///C:/b/abs_c4zmy2l696/croot/aiohttp_1670009573673/work
aioice==0.7.6
aiortc==1.4.0
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
altair==4.2.2
appdirs==1.4.4
astor==0.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout @ file:///C:/b/abs_43ozhz2a8g/croots/recipe/async-timeout_1664876362767/work
attrs @ file:///C:/b/abs_09s3y775ra/croot/attrs_1668696195628/work
av==10.0.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646038595831/work
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
datasets==2.10.1
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
diffusers==0.14.0
dill==0.3.6
dnspython==2.3.0
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.10.0
Flask==2.2.3
flatbuffers @ file:///tmp/build/80754af9/python-flatbuffers_1614345733764/work
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
frozenlist @ file:///C:/b/abs_2bb5uzghsi/croot/frozenlist_1670004511812/work
fsspec==2023.3.0
ftfy==6.1.1
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
gitdb==4.0.10
GitPython==3.1.31
google-auth==1.35.0
google-auth-oauthlib==0.4.1
google-crc32c==1.5.0
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///C:/ci/grpcio_1637590978642/work
h5py @ file:///C:/ci/h5py_1659089875384/work
huggingface-hub==0.13.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562621412/work
ipykernel @ file:///C:/b/abs_b4f07tbsyd/croot/ipykernel_1672767104060/work
ipython @ file:///C:/b/abs_d3h279dv3h/croot/ipython_1676582236558/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib==1.2.0
jsonschema==4.17.3
jupyter_client @ file:///C:/b/abs_d8fk_kz9zk/croot/jupyter_client_1676330195659/work
jupyter_core @ file:///C:/b/abs_bd7elvu3w2/croot/jupyter_core_1676538600510/work
keras==2.12.0
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keybert==0.7.0
Markdown @ file:///C:/b/abs_98lv_ucina/croot/markdown_1671541919225/work
markdown-it-py==2.2.0
MarkupSafe @ file:///C:/ci/markupsafe_1654508077284/work
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
mpmath==1.2.1
multidict @ file:///C:/b/abs_6cx_8w3cv2/croot/multidict_1665674238352/work
multiprocess==0.70.14
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
netifaces==0.11.0
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
nltk==3.8.1
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
oauthlib @ file:///C:/b/abs_2eoymqc2ow/croot/oauthlib_1665490906043/work
openai==0.27.2
opencv-python==4.7.0.72
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
pandas==1.5.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.17.2
psutil==5.9.4
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==11.0.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
PyJWT @ file:///C:/ci/pyjwt_1657511236979/work
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyrsistent==0.19.3
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
regex==2022.10.31
requests @ file:///C:/b/abs_80_edxtvhu/croot/requests_1678709752318/work
requests-oauthlib==1.3.0
responses==0.18.0
rich==13.3.2
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn==1.2.2
scipy==1.10.0
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.20.0
streamlit-webrtc==0.45.0
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/b/abs_2fhvpo862s/croot/tensorboard-data-server_1670853600144/work/tensorboard_data_server-0.6.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.6.0
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor @ file:///C:/b/abs_16qe7jhw7n/croot/termcolor_1668084642458/work
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado @ file:///C:/ci/tornado_1662458743919/work
tqdm==4.65.0
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
transformers==4.27.1
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
tzdata==2022.7
tzlocal==4.2
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.6
Werkzeug @ file:///C:/b/abs_17q5kgb8bo/croot/werkzeug_1671216014857/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt @ file:///C:/Windows/Temp/abs_7c3dd407-1390-477a-b542-fd15df6a24085_diwiza/croots/recipe/wrapt_1657814452175/work
xxhash==3.2.0
yarl @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e5nlunspj6/croots/recipe/yarl_1661437086516/work
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
